\name{helperFunctions}
\alias{getChangedIdx}

\title{arraysmooth Helper functions}

\description{
   arraysmooth Helper functions
}

\usage{
  getChangedIdx(changed,up)
}

\arguments{
   \item{changed}{}
   \item{up}{}
}

\details{
  Should not be called directly
}
\value{
}
\author{Jan Oosting}

\keyword{internal}
